Fix binding when vhost / is used
This commit fixes a bug when a vhost of / is used for a binding.
The vhost / clashed with the internal ID separator which is also /.

The fix checks the length of the internal ID when split and if it's
a certain length, normalizes the slice to the correct length by
removing the first element of the slice.