Allow integer values for policy definitions
This commit fixes a bug where policy definitions with integer
values were not working correctly due to their implicit cast
to a string.

This commit fixes it by detecting if a value is an integer and
setting it correctly in the API request body. Converting back
to a string when the value is read is also accounted for.