Correctly handle "not found" errors
Probably during the rabbithole update from 7235f3, the errors has currently
checked by the provider don't correctly detect that a resource has gone
missing on the cluster.

This creates unexpected "404 Not Found" errors at the end of any
refresh/plan/apply/destroy if any of the resources previously created by
Terraform (and known in the Terraform state) has been removed from the
RabbitMQ cluster, outside of Terraform.

I only tested a few, but this should affect all the resources: queues,
exchanges, vhost, permissions, users and policies (verified only on
vhosts, permissions, users and policies).

In this case, the RabbitMQ provider tries to read the resource it created
in the past, receives a 404 response from the cluster but doesn't treat it
correctly and returns a generic response "404 Object Not Found".

This change tries to read the error as a rabbithole error and handle the
404 response correctly.