Fix update ordering and deletion
Switch to `schema.TypeSet` for `permissions`; which as well as
being correct; fixes issues where the terraform order and RabbitMQ
order differ.

Rabbit-hole currently only supports deleting all permissions for
a vhost/user. So to do an update we delete all, then recreate the
existing and new ones. This is not ideal as it could lead to
removing permisions if you try and add an invalid one.